# Conflicts:
#	features/content/product_preview/src/main/java/com/tokopedia/content/product/preview/utils/Constants.kt
#	features/content/product_preview/src/main/java/com/tokopedia/content/product/preview/view/activity/ProductPreviewActivity.kt
#	features/content/product_preview/src/main/java/com/tokopedia/content/product/preview/view/fragment/ProductFragment.kt
#	features/content/product_preview/src/main/java/com/tokopedia/content/product/preview/view/fragment/ProductPreviewFragment.kt
#	features/content/product_preview/src/main/java/com/tokopedia/content/product/preview/view/fragment/ReviewFragment.kt
#	features/content/product_preview/src/main/java/com/tokopedia/content/product/preview/viewmodel/ProductPreviewViewModel.kt
#	features/content/product_preview/src/main/java/com/tokopedia/content/product/preview/viewmodel/action/ProductPreviewUiAction.kt

+++ features/content/product_preview/src/main/java/com/tokopedia/content/product/preview/data/GetMiniProductInfoResponse.kt
@@ -41,9 +41,9 @@ data class GetMiniProductInfoResponse(
         @SerializedName("isActive")
         val isActive: Boolean = false,
         @SerializedName("discountPercentage")
-        val discountPercentage: Float = 0f,
+        val discountPercentage: Int = 0,
         @SerializedName("discountedPrice")
-        val discountedPrice: Float = 0f,
+        val discountedPrice: Double = 0.0,
     )
 
     data class Shop(
